% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-jsm2024.Rmd.
\docType{data}
\name{authors}
\alias{authors}
\title{JSM author names}
\format{
A character vector of 12,628 names from past and present JSMs
}
\source{
This data was from the JSM 2022-2024 programs.
}
\usage{
authors
}
\description{
JSM author names
}
\keyword{datasets}