% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-jsm2024.Rmd.
\docType{data}
\name{authors2024}
\alias{authors2024}
\title{JSM 2024 author names}
\format{
A character vector of 5,820 JSM author names
}
\source{
This data was from the JSM 2024 program.
}
\usage{
authors2024
}
\description{
JSM 2024 author names
}
\keyword{datasets}